Move OTP logic into new file
