+++ bot.py
@@ -1,12 +1,8 @@
 import os
-import re
-import sys
 import time
-import secrets
 import logging
 import sqlite3
 import discord
-import requests
 from discord import app_commands
 from discord.ext import commands
 from export import export_db_to_csv, import_csv_to_db
@@ -14,6 +10,7 @@ from datetime import datetime
 from zoneinfo import ZoneInfo
 
 import config
+from otp import generate_otp, send_email_otp, valid_email_domain
 from utils import log_admin, safe_guild_filename, safe_guild_name
 
 os.makedirs("logs", exist_ok=True)
@@ -62,40 +59,6 @@ def get_guild_db(guild: discord.Guild):
 # Active OTP dictionary
 pending_verifications = {}
 # (guild_id, user_id): {code, expires, last_sent, email}
-
-
-def generate_otp():
-    return "".join(secrets.token_hex(nbytes=config.OTP_LENGTH // 2).upper())
-
-
-def valid_email_domain(email):
-    match = re.match(r"[^@]+@([^@]+\.[^@]+)", email)
-    if not match:
-        return False
-    domain = match.group(1).lower()
-    return domain in config.ALLOWED_DOMAINS
-
-
-def send_email_otp(to_email, code):
-    if not config.MAILGUN_API_KEY:
-        print(f"OTP for {to_email}: {code}")
-
-        class MockResponse:
-            status_code = 200
-
-        return MockResponse()
-
-    return requests.post(
-        f"https://api.mailgun.net/v3/{config.MAILGUN_DOMAIN}/messages",
-        auth=("api", config.MAILGUN_API_KEY),
-        data={
-            "from": config.MAILGUN_FROM,
-            "to": [to_email],
-            "subject": "Verify your email address",
-            "text": f"Your verification code is: {code}\nExpires in {config.OTP_EXPIRY_SECONDS/60} minutes.",
-        },
-        timeout=10,
-    )
 
 
 # create a popup in discord upon /verify invocation
