+++ utils.py
@@ -1,5 +1,6 @@
 import os
 import re
+import sys
 import discord
 import config
